Refactored associations views into multiple files
