+++ associations/views/__init__.py
@@ -1,0 +1,7 @@
+from associations.views.association import *
+from associations.views.events import *
+from associations.views.library import *
+from associations.views.marketplace import *
+from associations.views.news import *
+from associations.views.page import *
+from associations.views.vote import *
