+++ associations/views/news.py
@@ -1,0 +1,17 @@
+from rest_framework import viewsets
+
+from associations.models import News
+from associations.serializers import NewsSerializer
+
+
+class NewsViewSet(viewsets.ModelViewSet):
+    queryset = News.objects.all()
+    serializer_class = NewsSerializer
+
+    def get_queryset(self):
+        queryset = News.objects.all()
+        association_name = self.request.query_params.get('association', None)
+        if association_name is not None:
+            queryset = queryset.filter(association=association_name)
+        return queryset
+
