+++ associations/views/page.py
@@ -1,0 +1,20 @@
+from rest_framework import viewsets
+
+from associations.models import Page
+from associations.serializers import NewsSerializer
+
+
+class PageViewSet(viewsets.ModelViewSet):
+    queryset = Page.objects.all()
+    serializer_class = NewsSerializer
+
+    def get_queryset(self):
+        queryset = Page.objects.all()
+        association_name = self.request.query_params.get('association', None)
+        if association_name is not None:
+            queryset = queryset.filter(association=association_name)
+        return queryset
+
+
+
+
